Refactor project structure and Supabase config
