+++ .gitignore
@@ -1,0 +1,2 @@
+voca-2/js/config.js
+
