+++ voca-2/auth/auth-reset-password.html
@@ -250,14 +250,7 @@
     
     <!-- Supabase -->
     <script src="https://unpkg.com/@supabase/supabase-js@2"></script>
-    <script>
-        // Initialize Supabase client globally
-        const supabaseClient = supabase.createClient(
-            'https://tkiqqpqplbgrebcbsiop.supabase.co',
-            'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJzdXBhYmFzZSIsInJlZiI6InRraXFxcHFwbGJncmViY2JzaW9wIiwicm9sZSI6ImFub24iLCJpYXQiOjE3NDc0NzM3MzYsImV4cCI6MjA2MzA0OTczNn0.FJ0C0d4PYJTTUSL55tDroIpz5jcW9gvYX9eSKoB3Iz4'
-        );
-    </script>
     <!-- Custom JavaScript -->
-    <script src="auth-reset-password.js"></script>
+    <script type="module" src="auth-reset-password.js"></script>
 </body>
 </html>
