+++ voca-2/auth/auth-sign-up.html
@@ -1,0 +1,140 @@
+<!DOCTYPE html>
+<html lang="vi">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <meta name="description" content="Đăng ký tài khoản Study Notes để bắt đầu ghi chú và quản lý việc học tập của bạn">
+    <title>Đăng ký - Study Notes</title>
+    
+    <!-- Favicon -->
+    <link rel="icon" type="image/png" href="../assets/images/favicon.png">
+    
+    <!-- CSS -->
+    <link rel="stylesheet" href="auth-styles.css">
+    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.4.0/css/all.min.css" 
+          integrity="sha512-iecdLmaskl7CVkqkXNQ/ZH/XLlvWZOJyj7Yy7tcenmpD1ypASozpmT/E0iPtmFIB46ZmdtAc9eNBvH0H/ZpiBw==" 
+          crossorigin="anonymous" referrerpolicy="no-referrer" />
+    
+    <!-- Fonts -->
+    <link href="https://fonts.googleapis.com/css2?family=Inter:wght@300;400;500;600;700&display=swap" rel="stylesheet">
+    
+    <style>
+        /* Ensure the page takes full height */
+        html, body {
+            height: 100%;
+            margin: 0;
+            padding: 0;
+        }
+        
+        /* Smooth scrolling */
+        html {
+            scroll-behavior: smooth;
+        }
+    </style>
+</head>
+<body>
+    <div class="auth-container">
+        <div class="auth-card">
+            <div class="auth-header">
+                <a href="../main/index.html" class="back-button">
+                    <i class="fas fa-arrow-left"></i>
+                </a>
+                <h1>Study Notes</h1>
+            </div>
+
+            <div class="auth-content">
+                <div class="form-header">
+                    <h2>Tạo tài khoản mới</h2>
+                    <p>Điền thông tin để đăng ký</p>
+                </div>
+
+
+                <!-- Error Message -->
+                <div id="error-message" class="error-message"></div>
+
+                <!-- Sign Up Form -->
+                <form id="signupForm" class="auth-form" onsubmit="event.preventDefault();">
+                    <div class="form-group">
+                        <label for="fullName">Họ và tên <span class="required">*</span></label>
+                        <input type="text" id="fullName" name="fullName" class="form-control" 
+                               placeholder="Nhập họ và tên" required 
+                               autocomplete="name" autofocus>
+                    </div>
+
+                    <div class="form-group">
+                        <label for="email">Email <span class="required">*</span></label>
+                        <input type="email" id="email" name="email" class="form-control" 
+                               placeholder="Nhập địa chỉ email" required
+                               autocomplete="email">
+                    </div>
+
+                    
+                    <div class="form-group">
+                        <label for="password">Mật khẩu <span class="required">*</span></label>
+                        <div class="password-input">
+                            <input type="password" id="password" name="password" class="form-control" 
+                                   placeholder="Nhập mật khẩu (tối thiểu 6 ký tự)" required
+                                   minlength="6" autocomplete="new-password">
+                            <button type="button" class="toggle-password" aria-label="Hiển thị mật khẩu">
+                                <i class="fas fa-eye"></i>
+                            </button>
+                        </div>
+                        <small class="form-text text-muted">Mật khẩu phải có ít nhất 6 ký tự</small>
+                    </div>
+                    
+                    <div class="form-group">
+                        <label for="confirmPassword">Xác nhận mật khẩu <span class="required">*</span></label>
+                        <div class="password-input">
+                            <input type="password" id="confirmPassword" name="confirmPassword" class="form-control" 
+                                   placeholder="Nhập lại mật khẩu" required
+                                   minlength="6" autocomplete="new-password">
+                            <button type="button" class="toggle-password" aria-label="Hiển thị mật khẩu">
+                                <i class="fas fa-eye"></i>
+                            </button>
+                        </div>
+                    </div>
+                    
+                    <div class="form-group">
+                        <button type="submit" class="btn btn-primary btn-block">
+                            <span class="btn-text">Đăng ký</span>
+                            <span class="btn-loading">
+                                <i class="fas fa-spinner fa-spin"></i>
+                            </span>
+                        </button>
+                    </div>
+                    
+                    <div class="auth-divider">
+                        <span>hoặc</span>
+                    </div>
+                    
+                    <div class="social-login">
+                        <button type="button" class="btn btn-outline btn-google">
+                            <i class="fab fa-google"></i> Đăng ký với Google
+                        </button>
+                        <button type="button" class="btn btn-outline btn-facebook">
+                            <i class="fab fa-facebook-f"></i> Đăng ký với Facebook
+                        </button>
+                    </div>
+                </form>
+
+                <div class="auth-footer">
+                    <p>Đã có tài khoản? <a href="auth-sign-in.html">Đăng nhập ngay</a></p>
+                    <p class="terms">Bằng việc đăng ký, bạn đồng ý với <a href="../terms.html" target="_blank">Điều khoản dịch vụ</a> 
+                    và <a href="../privacy.html" target="_blank">Chính sách bảo mật</a> của chúng tôi.</p>
+                </div>
+            </div>
+        </div>
+    </div>
+
+    <!-- Loading Overlay -->
+    <div class="loading-overlay">
+        <div class="loading-spinner"></div>
+        <p class="loading-text">Đang xử lý...</p>
+    </div>
+
+    <!-- Scripts -->
+    <!-- Supabase -->
+    <script src="https://unpkg.com/@supabase/supabase-js@2"></script>
+    <script type="module" src="auth-sign-up.js"></script>
+</body>
+</html>
