+++ voca-2/js/config.example.js
@@ -1,0 +1,2 @@
+export const supabaseUrl = 'https://YOUR_SUPABASE_PROJECT.supabase.co';
+export const supabaseAnonKey = 'YOUR_SUPABASE_ANON_KEY';
